Improve test organization
This splits up the widget tests and gives a better description of what each
test file is testing.
