+++ test/background/views-test.js
@@ -8,7 +8,7 @@ import { expect } from "chai";
 
 import * as views from "../../src/webextension/background/views";
 
-describe("views", () => {
+describe("background > views", () => {
   beforeEach(() => {
     browser.tabs.mockClearTabs();
   });
